Rename all file from `.h` to `.hpp`
