+++ content/lab/l2/solution/fruit.hpp
