+++ content/lab/l2/solution/holey_string.hpp
