+++ content/lab/l2/solution/memory_manipulation.hpp
