+++ content/lab/l2/solution/vector3.hpp
