+++ content/lab/l2/src/holey_string.hpp
