+++ content/lab/l2/src/memory_manipulation.hpp
